refactor: create one file for all interface
